Add some more utility functions, and fix a typo in ALIGNED()
This adds a bunch of convenience macros, including a linked list
implementation derived from the kernel's, but without the
__list_add_valid() and __list_del_entry_valid() calls.

I think the idea of the __list_add_valid() and __list_del_entry_valid()
calls, which always return true, is to force the compiler to evaluate
things at an earlier pass.  With -Wanalyzer, the compiler notices that
these are called from a conditional and assumes it can return false,
even though returning true is *literally the only thing they do*, and
then it complains about the other path that could (not possibly) be
taken.  So we just don't need them here.

Signed-off-by: Peter Jones <pjones@redhat.com>